feat: add unittests for Allen example
- VoxelIntersection: add skips in case of absence of matching labels.
- add simple density file for tests
- add unittests
- move skip_tests for allen to bsb_test
- fix allen annotations download
- fix Readmes
